+++ changelog.md
@@ -8,6 +8,7 @@
 - added: comment user on top
 - added: hide post actions
 - added: background for sticky posts
+- added: infinite scrolling toggle
 - fixed: long press profile tab to switch accounts
 
 ## Build 58